+++ docs/CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [unreleased]
 
+### Added
+
+- Batch compute for both `compute()` and `compute_procedure()` methods
+- `FutureResultGroup` for batch computations
+
+### Change
+
+- Added `pydantic` `BaseModel` as base for `FutureResult` objects.
+
 ## [0.2.2]
 
 ### Added